+++ harness_designer/gl/canvas3d/focal_target.py
@@ -82,10 +82,10 @@ class FocalPoint3D(_generic.Generic3D):
             radius, 0.35, 19.0, 20.0, 30.0)))
 
         count = 2 * resolution * (resolution - 1) + 2
-        vertices = np.full((count, 3), [0.0, 0.0, 0.0], dtype=np.float64)
+        vertices = np.full((count, 3), [0.0, 0.0, 0.0], dtype=np.float32)
 
-        vertices[0] = np.array([0.0, 0.0, radius], dtype=np.float64)
-        vertices[1] = np.array([0.0, 0.0, -radius], dtype=np.float64)
+        vertices[0] = np.array([0.0, 0.0, radius], dtype=np.float32)
+        vertices[1] = np.array([0.0, 0.0, -radius], dtype=np.float32)
 
         step = math.pi / float(resolution)
 
@@ -103,7 +103,7 @@ class FocalPoint3D(_generic.Generic3D):
                 vertices[base + j] = np.array(
                     [alpha_sin * theta_cos,
                      alpha_sin * theta_sin,
-                     alpha_cos], dtype=np.float64) * radius
+                     alpha_cos], dtype=np.float32) * radius
 
         # Triangles for poles.
         faces = []
